fix: enable role based access control
