+++ backend/question-service/src/routes/question.routes.ts
@@ -1,4 +1,3 @@
-// import passport from 'passport'
 import {
     handleCreateQuestion,
     handleDeleteQuestion,
@@ -10,22 +9,20 @@ import {
     handleUpdateQuestion,
 } from '../controllers/question.controller'
 
-// import { Role } from '@repo/user-types'
+import { Role } from '@repo/user-types'
 import { Router } from 'express'
-// import { handleRoleBasedAccessControl } from '../middlewares/accessControl.middleware'
+import { handleRoleBasedAccessControl } from '../middlewares/accessControl.middleware'
+import passport from 'passport'
 
 const router = Router()
 
-// router.use(passport.authenticate('jwt', { session: false }))
+router.use(passport.authenticate('jwt', { session: false }))
 
 router.get('/', handleGetPaginatedQuestions)
 router.get('/:id', handleGetQuestionById)
-// router.post('/', handleRoleBasedAccessControl([Role.ADMIN]), handleCreateQuestion)
-// router.put('/:id', handleRoleBasedAccessControl([Role.ADMIN]), handleUpdateQuestion)
-// router.delete('/:id', handleRoleBasedAccessControl([Role.ADMIN]), handleDeleteQuestion)
-router.post('/', handleCreateQuestion)
-router.put('/:id', handleUpdateQuestion)
-router.delete('/:id', handleDeleteQuestion)
+router.post('/', handleRoleBasedAccessControl([Role.ADMIN]), handleCreateQuestion)
+router.put('/:id', handleRoleBasedAccessControl([Role.ADMIN]), handleUpdateQuestion)
+router.delete('/:id', handleRoleBasedAccessControl([Role.ADMIN]), handleDeleteQuestion)
 router.get('/filters', handleGetFilters)
 router.get('/sorts', handleGetSorts)
 router.get('/categories', handleGetCategories)
