+++ frontend/services/axios-middleware.ts
@@ -43,18 +43,4 @@ api.interceptors.response.use(
     }
 )
 
-// Request interceptor for all axios calls
-api.interceptors.request.use(
-    (config) => {
-        const token = sessionStorage.getItem('accessToken')
-        if (token) {
-            config.headers['Authorization'] = `Bearer ${token}`
-        }
-        return config
-    },
-    (error) => {
-        return Promise.reject(error)
-    }
-)
-
 export default api
