+++ frontend/services/axios-middleware2.ts
@@ -4,6 +4,20 @@ const api = axios.create({
     baseURL: 'http://localhost:3004',
 })
 
+// Request interceptor for all axios calls
+api.interceptors.request.use(
+    (config) => {
+        const token = sessionStorage.getItem('accessToken')
+        if (token) {
+            config.headers['Authorization'] = `Bearer ${token}`
+        }
+        return config
+    },
+    (error) => {
+        return Promise.reject(error)
+    }
+)
+
 // Response interceptor for all axios calls
 api.interceptors.response.use(
     (response) => {
